Expose a basic linter configs in the code editor component
