+++ apps/webapp/app/components/code/codeMirrorSetup.ts
@@ -1,14 +1,12 @@
 import { closeBrackets } from "@codemirror/autocomplete";
 import { indentWithTab } from "@codemirror/commands";
-import { jsonParseLinter } from "@codemirror/lang-json";
 import { bracketMatching } from "@codemirror/language";
-import { type Diagnostic, linter, lintGutter, lintKeymap } from "@codemirror/lint";
+import { lintKeymap } from "@codemirror/lint";
 import { highlightSelectionMatches } from "@codemirror/search";
 import { Prec, type Extension } from "@codemirror/state";
 import {
   drawSelection,
   dropCursor,
-  type EditorView,
   highlightActiveLine,
   highlightActiveLineGutter,
   highlightSpecialChars,
@@ -16,27 +14,12 @@ import {
   lineNumbers,
 } from "@codemirror/view";
 
-function emptyAwareJsonLinter() {
-  return (view: EditorView): Diagnostic[] => {
-    const content = view.state.doc.toString().trim();
-
-    // return no errors if content is empty
-    if (!content) {
-      return [];
-    }
-
-    return jsonParseLinter()(view);
-  };
-}
-
 export function getEditorSetup(showLineNumbers = true, showHighlights = true): Array<Extension> {
   const options = [
     drawSelection(),
     dropCursor(),
     bracketMatching(),
     closeBrackets(),
-    lintGutter(),
-    linter(emptyAwareJsonLinter()),
     Prec.highest(
       keymap.of([
         {
